move: Overview 디렉토리 생성하여 Overview, PropertyInfo 이동
components에 Overview 디렉토리 생성하고 Overview.tsx와
PropertyInfo.tsx를 이곳으로 이동.
매물 요약 관련 파일을 관리하는 디렉토리로.
